small changes : unit backend test  jobs ci,

+++ FrontEnd/Playwright/tests/setup/setup.edge.ts
@@ -12,13 +12,14 @@ setup('Edge-specific authentication', async ({ userProfilePage, loginPage, page,
     console.log(' Setting up Edge browser...');
     
     // Ensure directory exists
-    fs.mkdirSync(path.dirname(authFile), { recursive: true });
-    
-    // if file already exists remove old state of that file
-    if (fs.existsSync(authFile)) {
-        fs.unlinkSync(authFile);
-        return;
+    const dir = path.dirname(authFile);
+    if (!fs.existsSync(dir)) {
+        fs.mkdirSync(dir, { recursive: true });
     }
+       
+    // if file already exists remove old state of that file
+    if (fs.existsSync(authFile)) { fs.unlinkSync(authFile); }
+    if (fs.existsSync(dataFile)) { fs.unlinkSync(dataFile); }
     
     const id = Date.now();
     const userInputs = {
